Extracted alg2 functions to extra module.
